note: move sched_ramnote_* to noteram_driver.c
Let sched_note keep only common functions
The specific data output function is placed in the driver

Modifications:
Move sched_ramnote_* functions to ramnote_drivere.c
Change sched_note_flatten to a public function
Change note_common to a public function
Change sched_note_add to a private function
